refactory: The Home Component is divided into footer, header, contents components.
