+++ src/components/Contents.tsx
@@ -1,0 +1,9 @@
+const Contents = () => {
+  return (
+    <div>
+      <span>test</span>
+    </div>
+  );
+};
+
+export default Contents;
