+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as Footer } from './Footer';
+export { default as Header } from './Header';
